phase calculation by particles, trajectory tracking as a calculator
Describing and returning the phase of each particle is now up to the particle class.
The phases are collected at the end of the simulation as before, but also at each detector
hit. This allows us to simulate the particles beyond one detector and get a full description
of the phase space that each detector "saw".

Trajectory tracking is not done as a specific case, but just as one case of a concrete
(trivial) Calculator implementation.