Flatter package hierarchy, documentation improvements
This is a major restructuring in the top-level structure of the cminject package. The central point is the removal of the "definitions" subpackage, and the moving of all subpackages/modules below that module into the top-level "cminject" module. Furthermore, the "base" definition of each kind of object was moved to cminject.base. All this makes imports inside the software a lot nicer to read, write and use (IMO this is self-evident when comparing import sections before and after).