MF UpdateTemperature: only update temp if difference is non-negligible
The check is necessary for reasons of numerical stability. When the
temperature of the particle and the field are too close to each other,
the update starts diverging for numerical reasons, which we must of course
avoid.

We use np.isclose to do this check, using its default options. Changes might
be required in very specific cases where minute temperature differences are
important (but then, perhaps the units or scales should be changed
to avoid instability).