(WIP) working direct particle property access after all :)
By being more careful about when to interpolate within the generated function
(which is now called `get_at`), and by force-realizing this function before
calling @btime for analysis, we can get to a zero-overhead state and write our
acceleration functions directly in terms of particle instances :)
Also, no more need to represent particles as tuples, it seems. The defined
struct is just as efficient after compilation, now.

Overall: <1ms for a single ODE simulation, which makes me hopeful that it will extrapolate
roughly to the regime I've seen in `CMInjectScript.jl`: ~10s for 10^4 particles.